moved test to integration
