+++ test/integration/build_requires/build_requires_test.py
@@ -595,4 +595,39 @@ def test_requirement_in_wrong_method():
                 self.requires("foo/1.0")
         """)})
     tc.run('create . -cc="core:warnings_as_errors=[\'*\']"', assert_error=True)
-    assert "ERROR: deprecated: Requirements should only be added in the requirements()/build_requirements() methods, not configure()/config_options(), which might raise errors in the future." in tc.out
+    assert ("ERROR: deprecated: Requirements should only be added in the requirements()/"
+            "build_requirements() methods, not configure()/config_options(), which might "
+            "raise errors in the future.") in tc.out
+
+
+def test_transitive_build_scripts_error():
+    # https://github.com/conan-io/conan/issues/18235
+    c = TestClient()
+    meta = textwrap.dedent("""
+        from conan import ConanFile
+
+        class Meta(ConanFile):
+            name = "meta"
+            version = "0.1"
+            package_type = "build-scripts"
+            def requirements(self):
+                self.requires("dep/0.1")
+        """)
+    product = textwrap.dedent("""
+        from conan import ConanFile
+
+        class Product(ConanFile):
+            name = "product"
+            version = "0.1"
+            package_type = "build-scripts"
+            def requirements(self):
+                self.requires("meta/0.1")
+        """)
+
+    c.save({"dep/conanfile.py": GenConanfile("dep", "0.1"),
+            "meta/conanfile.py": meta,
+            "product/conanfile.py": product})
+    c.run("create dep")
+    c.run("create meta")
+    c.run("install product")
+    # It doesn't crash
